app: code split to User file in views folder
