+++ insurance-application/policyholder_app/src/views/User.css
@@ -1,0 +1,36 @@
+* {
+    box-sizing: border-box;
+    margin: 0;
+    padding: 0;
+  }
+  
+  body {
+    font-family: Arial, Helvetica, sans-serif;
+    line-height: 1.4;
+    background: #50b5b6;
+    font-family: monospace;
+    font-size: 120%
+  }
+  
+  a {
+    color: rgb(0, 0, 0);
+    text-decoration: none;
+    font-family: monospace;
+  }
+  
+  .container {
+    padding: 0 1rem;
+  }
+  
+  .btn {
+    display: inline-block;
+    border: none;
+    background: #555;
+    color: #fff;
+    padding: 7px 20px;
+    cursor: pointer;
+  }
+  
+  .btn:hover {
+    background: #666;
+  }
